Move `FloatEqualityWithoutAbs` into `Operators` lint pass
