+++ clippy_lints/src/lib.register_lints.rs
@@ -151,7 +151,6 @@ store.register_lints(&[
     exit::EXIT,
     explicit_write::EXPLICIT_WRITE,
     fallible_impl_from::FALLIBLE_IMPL_FROM,
-    float_equality_without_abs::FLOAT_EQUALITY_WITHOUT_ABS,
     float_literal::EXCESSIVE_PRECISION,
     float_literal::LOSSY_FLOAT_LITERAL,
     floating_point_arithmetic::IMPRECISE_FLOPS,
@@ -432,6 +431,7 @@ store.register_lints(&[
     operators::EQ_OP,
     operators::ERASING_OP,
     operators::FLOAT_ARITHMETIC,
+    operators::FLOAT_EQUALITY_WITHOUT_ABS,
     operators::INEFFECTIVE_BIT_MASK,
     operators::INTEGER_ARITHMETIC,
     operators::MISREFACTORED_ASSIGN_OP,
